requirements.txt
streamlit
pandas
pydeck
snowflake-connector-python
